refactor(RacingCarGame, Car): domain 패키지 생성
