+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.domain.RacingCarGame;
+
 public class Application {
     public static void main(String[] args) {
         RacingCarGame racingCarGame = new RacingCarGame();
